Make the paginated posts page work
