--- Sources/Blog/Components/Wrapper.swift
@@ -1,11 +1,0 @@
-import Plot
-import Publish
-
-struct Wrapper: ComponentContainer {
-    @ComponentBuilder var content: ContentProvider
-
-    var body: Component {
-        Div(content: content).class("wrapper")
-    }
-}
-
