+++ Sources/Blog/Constants.swift
@@ -1,4 +1,4 @@
 struct Constants {
-    static let numberOfItemsPerIndexPage: Int = 7
-    static let numberOfItemsPerTagsPage: Int = 7
+    static let numberOfItemsPerIndexPage: Int = 8
+    static let numberOfItemsPerTagsPage: Int = 8
 }
