+++ Sources/Blog/VaporBlogTheme.swift
@@ -44,11 +44,15 @@ private struct VaporBlogThemeHTMLFactory: HTMLFactory {
     func makePageHTML(for page: Page,
                       context: PublishingContext<Site>) throws -> HTML {
         let body: Node<HTML.DocumentContext> = .body {
-            SiteNavigation(context: context, selectedSelectionID: nil, currentSite: .blog, currentMainSitePage: nil)
-            Div {
+            if page.title == "Error 404 - Page not found" {
+                SiteNavigation(context: context, selectedSelectionID: nil, currentSite: .blog, currentMainSitePage: nil)
+                Div {
+                    page.body
+                }.class("container vapor-container")
+                SiteFooter(currentSite: .blog)
+            } else {
                 page.body
-            }.class("container vapor-container")
-            SiteFooter(currentSite: .blog)
+            }
         }
         let builder = VaporDesign<Site>(siteLanguage: context.site.language)
         return builder.buildHTML(for: page, context: context, body: body)
@@ -61,7 +65,7 @@ private struct VaporBlogThemeHTMLFactory: HTMLFactory {
             buildHead(for: page, context: context),
             .body {
                 SiteHeader(context: context, selectedSelectionID: nil)
-                Wrapper {
+                ComponentGroup {
                     TagsPage(
                         selectedTag: nil,
                         pageNumber: 1,
@@ -81,7 +85,7 @@ private struct VaporBlogThemeHTMLFactory: HTMLFactory {
             buildHead(for: page, context: context),
             .body {
                 SiteHeader(context: context, selectedSelectionID: nil)
-                Wrapper {
+                ComponentGroup {
                     TagsPage(
                         selectedTag: page.tag,
                         pageNumber: 1,
@@ -95,34 +99,17 @@ private struct VaporBlogThemeHTMLFactory: HTMLFactory {
     }
     
     func buildIndexPage(page: Location, context: PublishingContext<Site>) -> HTML {
-        let currentSite: CurrentSite = .blog
         let body: Node<HTML.DocumentContext> = .body {
-            SiteNavigation(context: context, selectedSelectionID: nil, currentSite: .blog, currentMainSitePage: nil)
-            Div {
-                H1("Articles, tools & resources for Vapor devs").class("vapor-blog-page-heading")
-                
-                Div {
-                    for item in context.paginatedItems.first ?? [] {
-                        Div {
-                            #warning("Fix")
-                            let authorImageURL = "https://design.vapor.codes/images/author-image-placeholder.png"
-                            let publishDate = DateFormatter.short.string(from: item.date)
-                            let blogPostData = BlogPostExtraData(length: "\(item.readingTime.minutes) minutes read", author: .init(name: item.metadata.author, imageURL: authorImageURL), publishedDate: publishDate)
-                            BlogCard(blogPostData: blogPostData, item: item, site: context.site)
-                        }.class("col")
-                    }
-                }.class("row row-cols-1 row-cols-lg-2 g-4 mb-5")
-                
-                Pagination(activePage: 1, numberOfPages: context.paginatedItems.count, pageURL: { pageNumber in
-                    context.index.paginatedPath(pageIndex: pageNumber - 1).absoluteString
-                })
-            }.class("container blog-container")
-            SiteFooter(currentSite: currentSite)
+            IndexPage(pageNumber: 1, context: context)
         }
         
         let builder = VaporDesign<Site>(siteLanguage: context.site.language)
         return builder.buildHTML(for: page, context: context, body: body)
     }
+    
+//    func buildTagPage(page: Location, context: PublishingContext<Site>) -> HTML {
+//        
+//    }
     
     func buildHead(for page: Location, context: PublishingContext<Blog>) -> Node<HTML.DocumentContext> {
             .head(for: page, on: context.site, stylesheetPaths: [
